Added cset verbose flag & default verbosity toggle command to report block state on cset
If cset is run with verbose=True, the new block state will be displayed as a result of the coommand.

cset default verbosity can be toggled on or off via "toggle_cset_verbose".  The verbose flag of each individual call will take precedence over the default verbosity setting (e.g. verbose is toggled on, but command is called with verbose=False, results will not be displayed).